@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: psdb/data/v1/data.proto
 
@@ -102,7 +102,7 @@ func file_psdb_data_v1_data_proto_rawDescGZIP() []byte {
 }
 
 var file_psdb_data_v1_data_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_psdb_data_v1_data_proto_goTypes = []interface{}{
+var file_psdb_data_v1_data_proto_goTypes = []any{
 	(Role)(0), // 0: psdb.data.v1.Role
 }
 var file_psdb_data_v1_data_proto_depIdxs = []int32{
