@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: psdb/v1alpha1/database.proto
 
@@ -8,9 +8,9 @@ package psdbv1alpha1
 
 import (
 	v1 "github.com/planetscale/psdb/types/psdb/data/v1"
-	v16 "github.com/planetscale/vitess-types/gen/vitess/query/v16"
-	v162 "github.com/planetscale/vitess-types/gen/vitess/vtgate/v16"
-	v161 "github.com/planetscale/vitess-types/gen/vitess/vtrpc/v16"
+	v19 "github.com/planetscale/vitess-types/gen/vitess/query/v19"
+	v192 "github.com/planetscale/vitess-types/gen/vitess/vtgate/v19"
+	v191 "github.com/planetscale/vitess-types/gen/vitess/vtrpc/v19"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
@@ -195,7 +195,7 @@ type ExecuteRequest struct {
 
 	Session       *Session                     `protobuf:"bytes,1,opt,name=session,proto3" json:"session,omitempty"`
 	Query         string                       `protobuf:"bytes,2,opt,name=query,proto3" json:"query,omitempty"`
-	BindVariables map[string]*v16.BindVariable `protobuf:"bytes,3,rep,name=bind_variables,json=bindVariables,proto3" json:"bind_variables,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	BindVariables map[string]*v19.BindVariable `protobuf:"bytes,3,rep,name=bind_variables,json=bindVariables,proto3" json:"bind_variables,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
 }
 
 func (x *ExecuteRequest) Reset() {
@@ -244,7 +244,7 @@ func (x *ExecuteRequest) GetQuery() string {
 	return ""
 }
 
-func (x *ExecuteRequest) GetBindVariables() map[string]*v16.BindVariable {
+func (x *ExecuteRequest) GetBindVariables() map[string]*v19.BindVariable {
 	if x != nil {
 		return x.BindVariables
 	}
@@ -257,8 +257,8 @@ type ExecuteResponse struct {
 	unknownFields protoimpl.UnknownFields
 
 	Session *Session         `protobuf:"bytes,1,opt,name=session,proto3" json:"session,omitempty"`
-	Result  *v16.QueryResult `protobuf:"bytes,2,opt,name=result,proto3" json:"result,omitempty"`
-	Error   *v161.RPCError   `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
+	Result  *v19.QueryResult `protobuf:"bytes,2,opt,name=result,proto3" json:"result,omitempty"`
+	Error   *v191.RPCError   `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
 	// optional timing, in seconds, for the response, if empty or 0, this wasn't set
 	// does not exist in the response from a StreamExecute.
 	Timing float64 `protobuf:"fixed64,4,opt,name=timing,proto3" json:"timing,omitempty"`
@@ -303,14 +303,14 @@ func (x *ExecuteResponse) GetSession() *Session {
 	return nil
 }
 
-func (x *ExecuteResponse) GetResult() *v16.QueryResult {
+func (x *ExecuteResponse) GetResult() *v19.QueryResult {
 	if x != nil {
 		return x.Result
 	}
 	return nil
 }
 
-func (x *ExecuteResponse) GetError() *v161.RPCError {
+func (x *ExecuteResponse) GetError() *v191.RPCError {
 	if x != nil {
 		return x.Error
 	}
@@ -331,7 +331,7 @@ type PrepareRequest struct {
 
 	Session       *Session                     `protobuf:"bytes,1,opt,name=session,proto3" json:"session,omitempty"`
 	Query         string                       `protobuf:"bytes,2,opt,name=query,proto3" json:"query,omitempty"`
-	BindVariables map[string]*v16.BindVariable `protobuf:"bytes,3,rep,name=bind_variables,json=bindVariables,proto3" json:"bind_variables,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	BindVariables map[string]*v19.BindVariable `protobuf:"bytes,3,rep,name=bind_variables,json=bindVariables,proto3" json:"bind_variables,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
 }
 
 func (x *PrepareRequest) Reset() {
@@ -380,7 +380,7 @@ func (x *PrepareRequest) GetQuery() string {
 	return ""
 }
 
-func (x *PrepareRequest) GetBindVariables() map[string]*v16.BindVariable {
+func (x *PrepareRequest) GetBindVariables() map[string]*v19.BindVariable {
 	if x != nil {
 		return x.BindVariables
 	}
@@ -393,8 +393,8 @@ type PrepareResponse struct {
 	unknownFields protoimpl.UnknownFields
 
 	Session *Session       `protobuf:"bytes,1,opt,name=session,proto3" json:"session,omitempty"`
-	Fields  []*v16.Field   `protobuf:"bytes,2,rep,name=fields,proto3" json:"fields,omitempty"`
-	Error   *v161.RPCError `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
+	Fields  []*v19.Field   `protobuf:"bytes,2,rep,name=fields,proto3" json:"fields,omitempty"`
+	Error   *v191.RPCError `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
 }
 
 func (x *PrepareResponse) Reset() {
@@ -436,14 +436,14 @@ func (x *PrepareResponse) GetSession() *Session {
 	return nil
 }
 
-func (x *PrepareResponse) GetFields() []*v16.Field {
+func (x *PrepareResponse) GetFields() []*v19.Field {
 	if x != nil {
 		return x.Fields
 	}
 	return nil
 }
 
-func (x *PrepareResponse) GetError() *v161.RPCError {
+func (x *PrepareResponse) GetError() *v191.RPCError {
 	if x != nil {
 		return x.Error
 	}
@@ -503,7 +503,7 @@ type CloseSessionResponse struct {
 	unknownFields protoimpl.UnknownFields
 
 	Session *Session       `protobuf:"bytes,1,opt,name=session,proto3" json:"session,omitempty"`
-	Error   *v161.RPCError `protobuf:"bytes,2,opt,name=error,proto3" json:"error,omitempty"`
+	Error   *v191.RPCError `protobuf:"bytes,2,opt,name=error,proto3" json:"error,omitempty"`
 }
 
 func (x *CloseSessionResponse) Reset() {
@@ -545,7 +545,7 @@ func (x *CloseSessionResponse) GetSession() *Session {
 	return nil
 }
 
-func (x *CloseSessionResponse) GetError() *v161.RPCError {
+func (x *CloseSessionResponse) GetError() *v191.RPCError {
 	if x != nil {
 		return x.Error
 	}
@@ -558,7 +558,7 @@ type Session struct {
 	unknownFields protoimpl.UnknownFields
 
 	Signature     []byte        `protobuf:"bytes,1,opt,name=signature,proto3" json:"signature,omitempty"`
-	VitessSession *v162.Session `protobuf:"bytes,2,opt,name=vitess_session,json=vitessSession,proto3" json:"vitess_session,omitempty"`
+	VitessSession *v192.Session `protobuf:"bytes,2,opt,name=vitess_session,json=vitessSession,proto3" json:"vitess_session,omitempty"`
 }
 
 func (x *Session) Reset() {
@@ -600,7 +600,7 @@ func (x *Session) GetSignature() []byte {
 	return nil
 }
 
-func (x *Session) GetVitessSession() *v162.Session {
+func (x *Session) GetVitessSession() *v192.Session {
 	if x != nil {
 		return x.VitessSession
 	}
@@ -615,11 +615,11 @@ var file_psdb_v1alpha1_database_proto_rawDesc = []byte{
 	0x70, 0x73, 0x64, 0x62, 0x2e, 0x76, 0x31, 0x61, 0x6c, 0x70, 0x68, 0x61, 0x31, 0x1a, 0x17, 0x70,
 	0x73, 0x64, 0x62, 0x2f, 0x64, 0x61, 0x74, 0x61, 0x2f, 0x76, 0x31, 0x2f, 0x64, 0x61, 0x74, 0x61,
 	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x1c, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2f, 0x71,
-	0x75, 0x65, 0x72, 0x79, 0x2f, 0x76, 0x31, 0x36, 0x2f, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e, 0x70,
+	0x75, 0x65, 0x72, 0x79, 0x2f, 0x76, 0x31, 0x39, 0x2f, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e, 0x70,
 	0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x1e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2f, 0x76, 0x74, 0x67,
-	0x61, 0x74, 0x65, 0x2f, 0x76, 0x31, 0x36, 0x2f, 0x76, 0x74, 0x67, 0x61, 0x74, 0x65, 0x2e, 0x70,
+	0x61, 0x74, 0x65, 0x2f, 0x76, 0x31, 0x39, 0x2f, 0x76, 0x74, 0x67, 0x61, 0x74, 0x65, 0x2e, 0x70,
 	0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x1c, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2f, 0x76, 0x74, 0x72,
-	0x70, 0x63, 0x2f, 0x76, 0x31, 0x36, 0x2f, 0x76, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x70, 0x72, 0x6f,
+	0x70, 0x63, 0x2f, 0x76, 0x31, 0x39, 0x2f, 0x76, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x70, 0x72, 0x6f,
 	0x74, 0x6f, 0x22, 0x5e, 0x0a, 0x04, 0x55, 0x73, 0x65, 0x72, 0x12, 0x1a, 0x0a, 0x08, 0x75, 0x73,
 	0x65, 0x72, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x75, 0x73,
 	0x65, 0x72, 0x6e, 0x61, 0x6d, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x70, 0x73, 0x69, 0x64, 0x18, 0x02,
@@ -652,7 +652,7 @@ var file_psdb_v1alpha1_database_proto_rawDesc = []byte{
 	0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03,
 	0x6b, 0x65, 0x79, 0x12, 0x34, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x18, 0x02, 0x20, 0x01,
 	0x28, 0x0b, 0x32, 0x1e, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x71, 0x75, 0x65, 0x72,
-	0x79, 0x2e, 0x76, 0x31, 0x36, 0x2e, 0x42, 0x69, 0x6e, 0x64, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62,
+	0x79, 0x2e, 0x76, 0x31, 0x39, 0x2e, 0x42, 0x69, 0x6e, 0x64, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62,
 	0x6c, 0x65, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a, 0x02, 0x38, 0x01, 0x22, 0xc4, 0x01,
 	0x0a, 0x0f, 0x45, 0x78, 0x65, 0x63, 0x75, 0x74, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
 	0x65, 0x12, 0x30, 0x0a, 0x07, 0x73, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01,
@@ -660,10 +660,10 @@ var file_psdb_v1alpha1_database_proto_rawDesc = []byte{
 	0x61, 0x31, 0x2e, 0x53, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x52, 0x07, 0x73, 0x65, 0x73, 0x73,
 	0x69, 0x6f, 0x6e, 0x12, 0x35, 0x0a, 0x06, 0x72, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x18, 0x02, 0x20,
 	0x01, 0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x71, 0x75, 0x65,
-	0x72, 0x79, 0x2e, 0x76, 0x31, 0x36, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79, 0x52, 0x65, 0x73, 0x75,
+	0x72, 0x79, 0x2e, 0x76, 0x31, 0x39, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79, 0x52, 0x65, 0x73, 0x75,
 	0x6c, 0x74, 0x52, 0x06, 0x72, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x12, 0x30, 0x0a, 0x05, 0x65, 0x72,
 	0x72, 0x6f, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x76, 0x69, 0x74, 0x65,
-	0x73, 0x73, 0x2e, 0x76, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x31, 0x36, 0x2e, 0x52, 0x50, 0x43,
+	0x73, 0x73, 0x2e, 0x76, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x31, 0x39, 0x2e, 0x52, 0x50, 0x43,
 	0x45, 0x72, 0x72, 0x6f, 0x72, 0x52, 0x05, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x12, 0x16, 0x0a, 0x06,
 	0x74, 0x69, 0x6d, 0x69, 0x6e, 0x67, 0x18, 0x04, 0x20, 0x01, 0x28, 0x01, 0x52, 0x06, 0x74, 0x69,
 	0x6d, 0x69, 0x6e, 0x67, 0x22, 0x93, 0x02, 0x0a, 0x0e, 0x50, 0x72, 0x65, 0x70, 0x61, 0x72, 0x65,
@@ -682,7 +682,7 @@ var file_psdb_v1alpha1_database_proto_rawDesc = []byte{
 	0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79,
 	0x12, 0x34, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32,
 	0x1e, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e, 0x76,
-	0x31, 0x36, 0x2e, 0x42, 0x69, 0x6e, 0x64, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52,
+	0x31, 0x39, 0x2e, 0x42, 0x69, 0x6e, 0x64, 0x56, 0x61, 0x72, 0x69, 0x61, 0x62, 0x6c, 0x65, 0x52,
 	0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a, 0x02, 0x38, 0x01, 0x22, 0xa6, 0x01, 0x0a, 0x0f, 0x50,
 	0x72, 0x65, 0x70, 0x61, 0x72, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x30,
 	0x0a, 0x07, 0x73, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32,
@@ -690,10 +690,10 @@ var file_psdb_v1alpha1_database_proto_rawDesc = []byte{
 	0x53, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x52, 0x07, 0x73, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e,
 	0x12, 0x2f, 0x0a, 0x06, 0x66, 0x69, 0x65, 0x6c, 0x64, 0x73, 0x18, 0x02, 0x20, 0x03, 0x28, 0x0b,
 	0x32, 0x17, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e,
-	0x76, 0x31, 0x36, 0x2e, 0x46, 0x69, 0x65, 0x6c, 0x64, 0x52, 0x06, 0x66, 0x69, 0x65, 0x6c, 0x64,
+	0x76, 0x31, 0x39, 0x2e, 0x46, 0x69, 0x65, 0x6c, 0x64, 0x52, 0x06, 0x66, 0x69, 0x65, 0x6c, 0x64,
 	0x73, 0x12, 0x30, 0x0a, 0x05, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b,
 	0x32, 0x1a, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x76, 0x74, 0x72, 0x70, 0x63, 0x2e,
-	0x76, 0x31, 0x36, 0x2e, 0x52, 0x50, 0x43, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x52, 0x05, 0x65, 0x72,
+	0x76, 0x31, 0x39, 0x2e, 0x52, 0x50, 0x43, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x52, 0x05, 0x65, 0x72,
 	0x72, 0x6f, 0x72, 0x22, 0x47, 0x0a, 0x13, 0x43, 0x6c, 0x6f, 0x73, 0x65, 0x53, 0x65, 0x73, 0x73,
 	0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x30, 0x0a, 0x07, 0x73, 0x65,
 	0x73, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x70, 0x73,
@@ -705,13 +705,13 @@ var file_psdb_v1alpha1_database_proto_rawDesc = []byte{
 	0x6c, 0x70, 0x68, 0x61, 0x31, 0x2e, 0x53, 0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x52, 0x07, 0x73,
 	0x65, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x30, 0x0a, 0x05, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x18,
 	0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x76,
-	0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x31, 0x36, 0x2e, 0x52, 0x50, 0x43, 0x45, 0x72, 0x72, 0x6f,
+	0x74, 0x72, 0x70, 0x63, 0x2e, 0x76, 0x31, 0x39, 0x2e, 0x52, 0x50, 0x43, 0x45, 0x72, 0x72, 0x6f,
 	0x72, 0x52, 0x05, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x22, 0x6a, 0x0a, 0x07, 0x53, 0x65, 0x73, 0x73,
 	0x69, 0x6f, 0x6e, 0x12, 0x1c, 0x0a, 0x09, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72, 0x65,
 	0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x09, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75, 0x72,
 	0x65, 0x12, 0x41, 0x0a, 0x0e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x5f, 0x73, 0x65, 0x73, 0x73,
 	0x69, 0x6f, 0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x76, 0x69, 0x74, 0x65,
-	0x73, 0x73, 0x2e, 0x76, 0x74, 0x67, 0x61, 0x74, 0x65, 0x2e, 0x76, 0x31, 0x36, 0x2e, 0x53, 0x65,
+	0x73, 0x73, 0x2e, 0x76, 0x74, 0x67, 0x61, 0x74, 0x65, 0x2e, 0x76, 0x31, 0x39, 0x2e, 0x53, 0x65,
 	0x73, 0x73, 0x69, 0x6f, 0x6e, 0x52, 0x0d, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x53, 0x65, 0x73,
 	0x73, 0x69, 0x6f, 0x6e, 0x32, 0xaf, 0x03, 0x0a, 0x08, 0x44, 0x61, 0x74, 0x61, 0x62, 0x61, 0x73,
 	0x65, 0x12, 0x5c, 0x0a, 0x0d, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x53, 0x65, 0x73, 0x73, 0x69,
@@ -760,7 +760,7 @@ func file_psdb_v1alpha1_database_proto_rawDescGZIP() []byte {
 }
 
 var file_psdb_v1alpha1_database_proto_msgTypes = make([]protoimpl.MessageInfo, 12)
-var file_psdb_v1alpha1_database_proto_goTypes = []interface{}{
+var file_psdb_v1alpha1_database_proto_goTypes = []any{
 	(*User)(nil),                  // 0: psdb.v1alpha1.User
 	(*CreateSessionRequest)(nil),  // 1: psdb.v1alpha1.CreateSessionRequest
 	(*CreateSessionResponse)(nil), // 2: psdb.v1alpha1.CreateSessionResponse
@@ -774,11 +774,11 @@ var file_psdb_v1alpha1_database_proto_goTypes = []interface{}{
 	nil,                           // 10: psdb.v1alpha1.ExecuteRequest.BindVariablesEntry
 	nil,                           // 11: psdb.v1alpha1.PrepareRequest.BindVariablesEntry
 	(v1.Role)(0),                  // 12: psdb.data.v1.Role
-	(*v16.QueryResult)(nil),       // 13: vitess.query.v16.QueryResult
-	(*v161.RPCError)(nil),         // 14: vitess.vtrpc.v16.RPCError
-	(*v16.Field)(nil),             // 15: vitess.query.v16.Field
-	(*v162.Session)(nil),          // 16: vitess.vtgate.v16.Session
-	(*v16.BindVariable)(nil),      // 17: vitess.query.v16.BindVariable
+	(*v19.QueryResult)(nil),       // 13: vitess.query.v19.QueryResult
+	(*v191.RPCError)(nil),         // 14: vitess.vtrpc.v19.RPCError
+	(*v19.Field)(nil),             // 15: vitess.query.v19.Field
+	(*v192.Session)(nil),          // 16: vitess.vtgate.v19.Session
+	(*v19.BindVariable)(nil),      // 17: vitess.query.v19.BindVariable
 }
 var file_psdb_v1alpha1_database_proto_depIdxs = []int32{
 	12, // 0: psdb.v1alpha1.User.role:type_name -> psdb.data.v1.Role
@@ -787,19 +787,19 @@ var file_psdb_v1alpha1_database_proto_depIdxs = []int32{
 	9,  // 3: psdb.v1alpha1.ExecuteRequest.session:type_name -> psdb.v1alpha1.Session
 	10, // 4: psdb.v1alpha1.ExecuteRequest.bind_variables:type_name -> psdb.v1alpha1.ExecuteRequest.BindVariablesEntry
 	9,  // 5: psdb.v1alpha1.ExecuteResponse.session:type_name -> psdb.v1alpha1.Session
-	13, // 6: psdb.v1alpha1.ExecuteResponse.result:type_name -> vitess.query.v16.QueryResult
-	14, // 7: psdb.v1alpha1.ExecuteResponse.error:type_name -> vitess.vtrpc.v16.RPCError
+	13, // 6: psdb.v1alpha1.ExecuteResponse.result:type_name -> vitess.query.v19.QueryResult
+	14, // 7: psdb.v1alpha1.ExecuteResponse.error:type_name -> vitess.vtrpc.v19.RPCError
 	9,  // 8: psdb.v1alpha1.PrepareRequest.session:type_name -> psdb.v1alpha1.Session
 	11, // 9: psdb.v1alpha1.PrepareRequest.bind_variables:type_name -> psdb.v1alpha1.PrepareRequest.BindVariablesEntry
 	9,  // 10: psdb.v1alpha1.PrepareResponse.session:type_name -> psdb.v1alpha1.Session
-	15, // 11: psdb.v1alpha1.PrepareResponse.fields:type_name -> vitess.query.v16.Field
-	14, // 12: psdb.v1alpha1.PrepareResponse.error:type_name -> vitess.vtrpc.v16.RPCError
+	15, // 11: psdb.v1alpha1.PrepareResponse.fields:type_name -> vitess.query.v19.Field
+	14, // 12: psdb.v1alpha1.PrepareResponse.error:type_name -> vitess.vtrpc.v19.RPCError
 	9,  // 13: psdb.v1alpha1.CloseSessionRequest.session:type_name -> psdb.v1alpha1.Session
 	9,  // 14: psdb.v1alpha1.CloseSessionResponse.session:type_name -> psdb.v1alpha1.Session
-	14, // 15: psdb.v1alpha1.CloseSessionResponse.error:type_name -> vitess.vtrpc.v16.RPCError
-	16, // 16: psdb.v1alpha1.Session.vitess_session:type_name -> vitess.vtgate.v16.Session
-	17, // 17: psdb.v1alpha1.ExecuteRequest.BindVariablesEntry.value:type_name -> vitess.query.v16.BindVariable
-	17, // 18: psdb.v1alpha1.PrepareRequest.BindVariablesEntry.value:type_name -> vitess.query.v16.BindVariable
+	14, // 15: psdb.v1alpha1.CloseSessionResponse.error:type_name -> vitess.vtrpc.v19.RPCError
+	16, // 16: psdb.v1alpha1.Session.vitess_session:type_name -> vitess.vtgate.v19.Session
+	17, // 17: psdb.v1alpha1.ExecuteRequest.BindVariablesEntry.value:type_name -> vitess.query.v19.BindVariable
+	17, // 18: psdb.v1alpha1.PrepareRequest.BindVariablesEntry.value:type_name -> vitess.query.v19.BindVariable
 	1,  // 19: psdb.v1alpha1.Database.CreateSession:input_type -> psdb.v1alpha1.CreateSessionRequest
 	3,  // 20: psdb.v1alpha1.Database.Execute:input_type -> psdb.v1alpha1.ExecuteRequest
 	3,  // 21: psdb.v1alpha1.Database.StreamExecute:input_type -> psdb.v1alpha1.ExecuteRequest
@@ -823,7 +823,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_psdb_v1alpha1_database_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*User); i {
 			case 0:
 				return &v.state
@@ -835,7 +835,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 				return nil
 			}
 		}
-		file_psdb_v1alpha1_database_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateSessionRequest); i {
 			case 0:
 				return &v.state
@@ -847,7 +847,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 				return nil
 			}
 		}
-		file_psdb_v1alpha1_database_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateSessionResponse); i {
 			case 0:
 				return &v.state
@@ -859,7 +859,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 				return nil
 			}
 		}
-		file_psdb_v1alpha1_database_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ExecuteRequest); i {
 			case 0:
 				return &v.state
@@ -871,7 +871,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 				return nil
 			}
 		}
-		file_psdb_v1alpha1_database_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*ExecuteResponse); i {
 			case 0:
 				return &v.state
@@ -883,7 +883,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 				return nil
 			}
 		}
-		file_psdb_v1alpha1_database_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*PrepareRequest); i {
 			case 0:
 				return &v.state
@@ -895,7 +895,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 				return nil
 			}
 		}
-		file_psdb_v1alpha1_database_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*PrepareResponse); i {
 			case 0:
 				return &v.state
@@ -907,7 +907,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 				return nil
 			}
 		}
-		file_psdb_v1alpha1_database_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*CloseSessionRequest); i {
 			case 0:
 				return &v.state
@@ -919,7 +919,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 				return nil
 			}
 		}
-		file_psdb_v1alpha1_database_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*CloseSessionResponse); i {
 			case 0:
 				return &v.state
@@ -931,7 +931,7 @@ func file_psdb_v1alpha1_database_proto_init() {
 				return nil
 			}
 		}
-		file_psdb_v1alpha1_database_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_psdb_v1alpha1_database_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*Session); i {
 			case 0:
 				return &v.state
