@@ -1,14 +1,14 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: psdbconnect/v1alpha1/connect.proto
 
 package psdbconnectv1alpha1
 
 import (
-	v16 "github.com/planetscale/vitess-types/gen/vitess/query/v16"
-	v161 "github.com/planetscale/vitess-types/gen/vitess/vtrpc/v16"
+	v19 "github.com/planetscale/vitess-types/gen/vitess/query/v19"
+	v191 "github.com/planetscale/vitess-types/gen/vitess/vtrpc/v19"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
@@ -199,7 +199,7 @@ type DeletedRow struct {
 	unknownFields protoimpl.UnknownFields
 
 	// This result will contain only the primary keys from the deleted row.
-	Result *v16.QueryResult `protobuf:"bytes,1,opt,name=result,proto3" json:"result,omitempty"`
+	Result *v19.QueryResult `protobuf:"bytes,1,opt,name=result,proto3" json:"result,omitempty"`
 }
 
 func (x *DeletedRow) Reset() {
@@ -234,7 +234,7 @@ func (*DeletedRow) Descriptor() ([]byte, []int) {
 	return file_psdbconnect_v1alpha1_connect_proto_rawDescGZIP(), []int{1}
 }
 
-func (x *DeletedRow) GetResult() *v16.QueryResult {
+func (x *DeletedRow) GetResult() *v19.QueryResult {
 	if x != nil {
 		return x.Result
 	}
@@ -248,9 +248,9 @@ type UpdatedRow struct {
 	unknownFields protoimpl.UnknownFields
 
 	// All values of the table before the update was made.
-	Before *v16.QueryResult `protobuf:"bytes,1,opt,name=before,proto3" json:"before,omitempty"`
+	Before *v19.QueryResult `protobuf:"bytes,1,opt,name=before,proto3" json:"before,omitempty"`
 	// All values of the table ater the update was made.
-	After *v16.QueryResult `protobuf:"bytes,2,opt,name=after,proto3" json:"after,omitempty"`
+	After *v19.QueryResult `protobuf:"bytes,2,opt,name=after,proto3" json:"after,omitempty"`
 }
 
 func (x *UpdatedRow) Reset() {
@@ -285,14 +285,14 @@ func (*UpdatedRow) Descriptor() ([]byte, []int) {
 	return file_psdbconnect_v1alpha1_connect_proto_rawDescGZIP(), []int{2}
 }
 
-func (x *UpdatedRow) GetBefore() *v16.QueryResult {
+func (x *UpdatedRow) GetBefore() *v19.QueryResult {
 	if x != nil {
 		return x.Before
 	}
 	return nil
 }
 
-func (x *UpdatedRow) GetAfter() *v16.QueryResult {
+func (x *UpdatedRow) GetAfter() *v19.QueryResult {
 	if x != nil {
 		return x.After
 	}
@@ -306,11 +306,11 @@ type SyncResponse struct {
 	unknownFields protoimpl.UnknownFields
 
 	// An array of rows that denote inserts into the table.
-	Result []*v16.QueryResult `protobuf:"bytes,1,rep,name=result,proto3" json:"result,omitempty"`
+	Result []*v19.QueryResult `protobuf:"bytes,1,rep,name=result,proto3" json:"result,omitempty"`
 	// A state object to use that denotes the current state of the SyncResponse.
 	Cursor *TableCursor `protobuf:"bytes,2,opt,name=cursor,proto3" json:"cursor,omitempty"`
 	// Any errors encountered in streaming data from the table.
-	Error *v161.RPCError `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
+	Error *v191.RPCError `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
 	// An array of rows that denote deletes from the table.
 	Deletes []*DeletedRow `protobuf:"bytes,4,rep,name=deletes,proto3" json:"deletes,omitempty"`
 	// An array of rows that denote updates to the table.
@@ -349,7 +349,7 @@ func (*SyncResponse) Descriptor() ([]byte, []int) {
 	return file_psdbconnect_v1alpha1_connect_proto_rawDescGZIP(), []int{3}
 }
 
-func (x *SyncResponse) GetResult() []*v16.QueryResult {
+func (x *SyncResponse) GetResult() []*v19.QueryResult {
 	if x != nil {
 		return x.Result
 	}
@@ -363,7 +363,7 @@ func (x *SyncResponse) GetCursor() *TableCursor {
 	return nil
 }
 
-func (x *SyncResponse) GetError() *v161.RPCError {
+func (x *SyncResponse) GetError() *v191.RPCError {
 	if x != nil {
 		return x.Error
 	}
@@ -402,7 +402,7 @@ type TableCursor struct {
 	// the SyncRequest will fail.
 	Position string `protobuf:"bytes,3,opt,name=position,proto3" json:"position,omitempty"`
 	// Any known last known primary key values from the a previous session.
-	LastKnownPk *v16.QueryResult `protobuf:"bytes,4,opt,name=last_known_pk,json=lastKnownPk,proto3" json:"last_known_pk,omitempty"`
+	LastKnownPk *v19.QueryResult `protobuf:"bytes,4,opt,name=last_known_pk,json=lastKnownPk,proto3" json:"last_known_pk,omitempty"`
 }
 
 func (x *TableCursor) Reset() {
@@ -458,7 +458,7 @@ func (x *TableCursor) GetPosition() string {
 	return ""
 }
 
-func (x *TableCursor) GetLastKnownPk() *v16.QueryResult {
+func (x *TableCursor) GetLastKnownPk() *v19.QueryResult {
 	if x != nil {
 		return x.LastKnownPk
 	}
@@ -472,9 +472,9 @@ var file_psdbconnect_v1alpha1_connect_proto_rawDesc = []byte{
 	0x61, 0x6c, 0x70, 0x68, 0x61, 0x31, 0x2f, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x2e, 0x70,
 	0x72, 0x6f, 0x74, 0x6f, 0x12, 0x14, 0x70, 0x73, 0x64, 0x62, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63,
 	0x74, 0x2e, 0x76, 0x31, 0x61, 0x6c, 0x70, 0x68, 0x61, 0x31, 0x1a, 0x1c, 0x76, 0x69, 0x74, 0x65,
-	0x73, 0x73, 0x2f, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2f, 0x76, 0x31, 0x36, 0x2f, 0x71, 0x75, 0x65,
+	0x73, 0x73, 0x2f, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2f, 0x76, 0x31, 0x39, 0x2f, 0x71, 0x75, 0x65,
 	0x72, 0x79, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x1c, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73,
-	0x2f, 0x76, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x76, 0x31, 0x36, 0x2f, 0x76, 0x74, 0x72, 0x70, 0x63,
+	0x2f, 0x76, 0x74, 0x72, 0x70, 0x63, 0x2f, 0x76, 0x31, 0x39, 0x2f, 0x76, 0x74, 0x72, 0x70, 0x63,
 	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0xd5, 0x02, 0x0a, 0x0b, 0x53, 0x79, 0x6e, 0x63, 0x52,
 	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1d, 0x0a, 0x0a, 0x74, 0x61, 0x62, 0x6c, 0x65, 0x5f,
 	0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x74, 0x61, 0x62, 0x6c,
@@ -499,27 +499,27 @@ var file_psdbconnect_v1alpha1_connect_proto_rawDesc = []byte{
 	0x73, 0x18, 0x08, 0x20, 0x03, 0x28, 0x09, 0x52, 0x05, 0x63, 0x65, 0x6c, 0x6c, 0x73, 0x22, 0x43,
 	0x0a, 0x0a, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x64, 0x52, 0x6f, 0x77, 0x12, 0x35, 0x0a, 0x06,
 	0x72, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x76,
-	0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e, 0x76, 0x31, 0x36, 0x2e,
+	0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e, 0x76, 0x31, 0x39, 0x2e,
 	0x51, 0x75, 0x65, 0x72, 0x79, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x52, 0x06, 0x72, 0x65, 0x73,
 	0x75, 0x6c, 0x74, 0x22, 0x78, 0x0a, 0x0a, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x64, 0x52, 0x6f,
 	0x77, 0x12, 0x35, 0x0a, 0x06, 0x62, 0x65, 0x66, 0x6f, 0x72, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28,
 	0x0b, 0x32, 0x1d, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x71, 0x75, 0x65, 0x72, 0x79,
-	0x2e, 0x76, 0x31, 0x36, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74,
+	0x2e, 0x76, 0x31, 0x39, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74,
 	0x52, 0x06, 0x62, 0x65, 0x66, 0x6f, 0x72, 0x65, 0x12, 0x33, 0x0a, 0x05, 0x61, 0x66, 0x74, 0x65,
 	0x72, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73,
-	0x2e, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e, 0x76, 0x31, 0x36, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79,
+	0x2e, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e, 0x76, 0x31, 0x39, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79,
 	0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x52, 0x05, 0x61, 0x66, 0x74, 0x65, 0x72, 0x22, 0xaa, 0x02,
 	0x0a, 0x0c, 0x53, 0x79, 0x6e, 0x63, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x35,
 	0x0a, 0x06, 0x72, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x1d,
 	0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e, 0x76, 0x31,
-	0x36, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x52, 0x06, 0x72,
+	0x39, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x52, 0x06, 0x72,
 	0x65, 0x73, 0x75, 0x6c, 0x74, 0x12, 0x39, 0x0a, 0x06, 0x63, 0x75, 0x72, 0x73, 0x6f, 0x72, 0x18,
 	0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x21, 0x2e, 0x70, 0x73, 0x64, 0x62, 0x63, 0x6f, 0x6e, 0x6e,
 	0x65, 0x63, 0x74, 0x2e, 0x76, 0x31, 0x61, 0x6c, 0x70, 0x68, 0x61, 0x31, 0x2e, 0x54, 0x61, 0x62,
 	0x6c, 0x65, 0x43, 0x75, 0x72, 0x73, 0x6f, 0x72, 0x52, 0x06, 0x63, 0x75, 0x72, 0x73, 0x6f, 0x72,
 	0x12, 0x30, 0x0a, 0x05, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32,
 	0x1a, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x76, 0x74, 0x72, 0x70, 0x63, 0x2e, 0x76,
-	0x31, 0x36, 0x2e, 0x52, 0x50, 0x43, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x52, 0x05, 0x65, 0x72, 0x72,
+	0x31, 0x39, 0x2e, 0x52, 0x50, 0x43, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x52, 0x05, 0x65, 0x72, 0x72,
 	0x6f, 0x72, 0x12, 0x3a, 0x0a, 0x07, 0x64, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x73, 0x18, 0x04, 0x20,
 	0x03, 0x28, 0x0b, 0x32, 0x20, 0x2e, 0x70, 0x73, 0x64, 0x62, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63,
 	0x74, 0x2e, 0x76, 0x31, 0x61, 0x6c, 0x70, 0x68, 0x61, 0x31, 0x2e, 0x44, 0x65, 0x6c, 0x65, 0x74,
@@ -536,7 +536,7 @@ var file_psdbconnect_v1alpha1_connect_proto_rawDesc = []byte{
 	0x70, 0x6f, 0x73, 0x69, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x41, 0x0a, 0x0d, 0x6c, 0x61, 0x73, 0x74,
 	0x5f, 0x6b, 0x6e, 0x6f, 0x77, 0x6e, 0x5f, 0x70, 0x6b, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32,
 	0x1d, 0x2e, 0x76, 0x69, 0x74, 0x65, 0x73, 0x73, 0x2e, 0x71, 0x75, 0x65, 0x72, 0x79, 0x2e, 0x76,
-	0x31, 0x36, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x52, 0x0b,
+	0x31, 0x39, 0x2e, 0x51, 0x75, 0x65, 0x72, 0x79, 0x52, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x52, 0x0b,
 	0x6c, 0x61, 0x73, 0x74, 0x4b, 0x6e, 0x6f, 0x77, 0x6e, 0x50, 0x6b, 0x2a, 0x31, 0x0a, 0x0a, 0x54,
 	0x61, 0x62, 0x6c, 0x65, 0x74, 0x54, 0x79, 0x70, 0x65, 0x12, 0x0b, 0x0a, 0x07, 0x72, 0x65, 0x70,
 	0x6c, 0x69, 0x63, 0x61, 0x10, 0x00, 0x12, 0x0b, 0x0a, 0x07, 0x70, 0x72, 0x69, 0x6d, 0x61, 0x72,
@@ -569,28 +569,28 @@ func file_psdbconnect_v1alpha1_connect_proto_rawDescGZIP() []byte {
 
 var file_psdbconnect_v1alpha1_connect_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_psdbconnect_v1alpha1_connect_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
-var file_psdbconnect_v1alpha1_connect_proto_goTypes = []interface{}{
+var file_psdbconnect_v1alpha1_connect_proto_goTypes = []any{
 	(TabletType)(0),         // 0: psdbconnect.v1alpha1.TabletType
 	(*SyncRequest)(nil),     // 1: psdbconnect.v1alpha1.SyncRequest
 	(*DeletedRow)(nil),      // 2: psdbconnect.v1alpha1.DeletedRow
 	(*UpdatedRow)(nil),      // 3: psdbconnect.v1alpha1.UpdatedRow
 	(*SyncResponse)(nil),    // 4: psdbconnect.v1alpha1.SyncResponse
 	(*TableCursor)(nil),     // 5: psdbconnect.v1alpha1.TableCursor
-	(*v16.QueryResult)(nil), // 6: vitess.query.v16.QueryResult
-	(*v161.RPCError)(nil),   // 7: vitess.vtrpc.v16.RPCError
+	(*v19.QueryResult)(nil), // 6: vitess.query.v19.QueryResult
+	(*v191.RPCError)(nil),   // 7: vitess.vtrpc.v19.RPCError
 }
 var file_psdbconnect_v1alpha1_connect_proto_depIdxs = []int32{
 	5,  // 0: psdbconnect.v1alpha1.SyncRequest.cursor:type_name -> psdbconnect.v1alpha1.TableCursor
 	0,  // 1: psdbconnect.v1alpha1.SyncRequest.tablet_type:type_name -> psdbconnect.v1alpha1.TabletType
-	6,  // 2: psdbconnect.v1alpha1.DeletedRow.result:type_name -> vitess.query.v16.QueryResult
-	6,  // 3: psdbconnect.v1alpha1.UpdatedRow.before:type_name -> vitess.query.v16.QueryResult
-	6,  // 4: psdbconnect.v1alpha1.UpdatedRow.after:type_name -> vitess.query.v16.QueryResult
-	6,  // 5: psdbconnect.v1alpha1.SyncResponse.result:type_name -> vitess.query.v16.QueryResult
+	6,  // 2: psdbconnect.v1alpha1.DeletedRow.result:type_name -> vitess.query.v19.QueryResult
+	6,  // 3: psdbconnect.v1alpha1.UpdatedRow.before:type_name -> vitess.query.v19.QueryResult
+	6,  // 4: psdbconnect.v1alpha1.UpdatedRow.after:type_name -> vitess.query.v19.QueryResult
+	6,  // 5: psdbconnect.v1alpha1.SyncResponse.result:type_name -> vitess.query.v19.QueryResult
 	5,  // 6: psdbconnect.v1alpha1.SyncResponse.cursor:type_name -> psdbconnect.v1alpha1.TableCursor
-	7,  // 7: psdbconnect.v1alpha1.SyncResponse.error:type_name -> vitess.vtrpc.v16.RPCError
+	7,  // 7: psdbconnect.v1alpha1.SyncResponse.error:type_name -> vitess.vtrpc.v19.RPCError
 	2,  // 8: psdbconnect.v1alpha1.SyncResponse.deletes:type_name -> psdbconnect.v1alpha1.DeletedRow
 	3,  // 9: psdbconnect.v1alpha1.SyncResponse.updates:type_name -> psdbconnect.v1alpha1.UpdatedRow
-	6,  // 10: psdbconnect.v1alpha1.TableCursor.last_known_pk:type_name -> vitess.query.v16.QueryResult
+	6,  // 10: psdbconnect.v1alpha1.TableCursor.last_known_pk:type_name -> vitess.query.v19.QueryResult
 	1,  // 11: psdbconnect.v1alpha1.Connect.Sync:input_type -> psdbconnect.v1alpha1.SyncRequest
 	4,  // 12: psdbconnect.v1alpha1.Connect.Sync:output_type -> psdbconnect.v1alpha1.SyncResponse
 	12, // [12:13] is the sub-list for method output_type
@@ -606,7 +606,7 @@ func file_psdbconnect_v1alpha1_connect_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_psdbconnect_v1alpha1_connect_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_psdbconnect_v1alpha1_connect_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*SyncRequest); i {
 			case 0:
 				return &v.state
@@ -618,7 +618,7 @@ func file_psdbconnect_v1alpha1_connect_proto_init() {
 				return nil
 			}
 		}
-		file_psdbconnect_v1alpha1_connect_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_psdbconnect_v1alpha1_connect_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*DeletedRow); i {
 			case 0:
 				return &v.state
@@ -630,7 +630,7 @@ func file_psdbconnect_v1alpha1_connect_proto_init() {
 				return nil
 			}
 		}
-		file_psdbconnect_v1alpha1_connect_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_psdbconnect_v1alpha1_connect_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdatedRow); i {
 			case 0:
 				return &v.state
@@ -642,7 +642,7 @@ func file_psdbconnect_v1alpha1_connect_proto_init() {
 				return nil
 			}
 		}
-		file_psdbconnect_v1alpha1_connect_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_psdbconnect_v1alpha1_connect_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*SyncResponse); i {
 			case 0:
 				return &v.state
@@ -654,7 +654,7 @@ func file_psdbconnect_v1alpha1_connect_proto_init() {
 				return nil
 			}
 		}
-		file_psdbconnect_v1alpha1_connect_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_psdbconnect_v1alpha1_connect_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*TableCursor); i {
 			case 0:
 				return &v.state
